+++ package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "@tailwindcss/ui": "^0.7.2",
     "@types/d3-format": "^2.0.0",
+    "@types/react-router-dom": "^5.1.6",
     "@zakodium/components-update": "^1.4.0",
     "@zakodium/tailwind-config": "^2.0.1",
     "autoprefixer": "^10.0.4",
@@ -44,6 +45,7 @@
     "react-dropzone": "^11.2.4",
     "react-plot": "^0.2.0",
     "react-popper": "^2.2.4",
+    "react-router-dom": "~5.2.0",
     "react-scripts": "4.0.1",
     "typescript": "^4.1.2",
     "web-vitals": "^0.2.4"
