+++ src/components/Info.tsx
@@ -3,31 +3,12 @@ import { ndParse } from 'ndim-parser';
 import { Plot, LineSeries, XAxis, YAxis } from 'react-plot';
 import { formatPrefix } from 'd3-format';
 
+import MetaTable from './MetaTable';
+
 interface InfoProps {
   text: string;
 }
 
-interface MetaTableProps {
-  meta: Record<string, string>;
-}
-
-function MetaTable({ meta }: MetaTableProps) {
-  return (
-    <div className="p-5 m-2 shadow sm:rounded-lg">
-      <div className="text-2xl font-semibold text-blue-500">Metadata</div>
-      <table>
-        <tbody className="inline-block overflow-auto border-t border-gray-200 h-96">
-          {Object.keys(meta).map((key) => (
-            <tr key={key}>
-              <td className="font-medium">{key}</td>
-              <td>{meta[key]}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
 export default function Info({ text }: InfoProps) {
   const { data, meta } = ndParse(text);
